Separate out slice tree related functions and add comments explaining why we need slice trees
